emulator: lifecycle controller reads state from fuses
This makes OTP fuses the source of truth for lifecycle state.

The emulator now:
- Reads lifecycle state from OTP file fuses if they exist
- Falls back to generating fuse data from --device-security-state CLI arg
- Passes lifecycle fuse data to OTP for provisioning when blank
- Derives Caliptra security state from the resolved LC state index
- Creates LcCtrl from the same decoded/computed state

We extracted the lifecycle encoding/decodign code into a common crate
so both the emulator and hw-model can share it.

Fixes #839
