+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "common/flash-image",
     "common/mcu-mbox",
     "common/otp-digest",
+    "common/otp-lifecycle",
     "common/mctp-vdm",
     "common/pldm",
     "common/poll",
@@ -219,6 +220,7 @@ mcu-testing-common = { path = "common/testing" }
 mcu-tock-veer = { path = "runtime/kernel/veer" }
 mctp-vdm-common = { path = "common/mctp-vdm" }
 otp-digest = { path = "common/otp-digest" }
+otp-lifecycle = { path = "common/otp-lifecycle" }
 pldm-common = { path = "common/pldm"}
 pldm-fw-pkg = { path = "emulator/bmc/pldm-fw-pkg" }
 pldm-ua = { path = "emulator/bmc/pldm-ua"}
