+++ emulator/app/Cargo.toml
@@ -39,6 +39,7 @@ log.workspace = true
 lazy_static.workspace = true
 mcu-mbox-common.workspace = true
 mcu-testing-common.workspace = true
+otp-lifecycle.workspace = true
 p384.workspace = true
 pldm-common.workspace = true
 pldm-fw-pkg.workspace = true
