+++ emulator/periph/Cargo.toml
@@ -22,6 +22,7 @@ lazy_static.workspace = true
 mcu-testing-common.workspace = true
 num_enum.workspace = true
 otp-digest.workspace = true
+otp-lifecycle.workspace = true
 registers-generated.workspace = true
 semver.workspace = true
 serde_json.workspace = true
