+++ hw/model/Cargo.toml
@@ -40,6 +40,7 @@ mcu-config-fpga = { workspace = true, optional = true }
 mcu-config.workspace = true
 mcu-rom-common.workspace = true
 mcu-testing-common.workspace = true
+otp-lifecycle.workspace = true
 p384.workspace = true
 poll-common.workspace = true
 rand.workspace = true
