+++ hw/model/src/model_emulated.rs
@@ -197,7 +197,20 @@ impl McuHwModel for ModelEmulated {
                 .copy_from_slice(&mem);
         }
 
-        let lc = LcCtrl::new();
+        // Derive LC state from the provisioned OTP fuses (source of truth).
+        let lc_bytes = &otp_mem[fuses::LIFE_CYCLE_BYTE_OFFSET
+            ..fuses::LIFE_CYCLE_BYTE_OFFSET + fuses::LIFE_CYCLE_BYTE_SIZE];
+        let (lc_state_index, lc_transition_cnt) = if lc_bytes.iter().any(|&b| b != 0) {
+            let mem: [u8; otp_lifecycle::LIFECYCLE_MEM_SIZE] = lc_bytes
+                .try_into()
+                .expect("lifecycle partition size mismatch");
+            let (state_idx, count) = otp_lifecycle::lc_decode_memory(&mem)?;
+            (state_idx as u32, count as u32)
+        } else {
+            (0, 0)
+        };
+
+        let lc = LcCtrl::with_state(lc_state_index, lc_transition_cnt);
 
         let otp = Otp::new(
             &clock.clone(),
@@ -264,9 +277,21 @@ impl McuHwModel for ModelEmulated {
 
         emulator_periph::AxiCDMA::set_dma_ram(&mut dma_ctrl, dma_ram.clone());
 
+        // Map LC state to Caliptra device lifecycle per the Caliptra SS HW spec
+        // (caliptra-ss docs/CaliptraSSHardwareSpecification.md, LCC state table).
         let device_lifecycle: Option<String> = match params.lifecycle_controller_state {
             Some(LifecycleControllerState::Dev) => Some("manufacturing".into()),
-            Some(LifecycleControllerState::Raw) => Some("unprovisioned".into()),
+            Some(
+                LifecycleControllerState::TestUnlocked0
+                | LifecycleControllerState::TestUnlocked1
+                | LifecycleControllerState::TestUnlocked2
+                | LifecycleControllerState::TestUnlocked3
+                | LifecycleControllerState::TestUnlocked4
+                | LifecycleControllerState::TestUnlocked5
+                | LifecycleControllerState::TestUnlocked6
+                | LifecycleControllerState::TestUnlocked7,
+            ) => Some("unprovisioned".into()),
+            // Raw, TestLocked*, Prod, ProdEnd, Rma, Scrap all map to production
             _ => Some("production".into()),
         };
